Activity jumps: t_offset_sec -> t_offset_ms to avoid same-second collisions
Two jumps landing in the same whole second (e.g. block+attack) truncated
to the same t_offset_sec and insertOrIgnore silently dropped one. Storing
milliseconds preserves both. Chart JS divides by 1000 for the x-axis
(fmtClock still expects seconds). Coordinated with iOS: server must
deploy first, then iOS switches to sending t in ms.

+++ backend/app/Http/Controllers/Api/ActivityJumpController.php
@@ -20,7 +20,7 @@ class ActivityJumpController extends Controller
 
         $data = $request->validate([
             'jumps'              => ['required', 'array', 'max:5000'],
-            'jumps.*.t'          => ['required', 'numeric', 'min:0'],
+            'jumps.*.t'          => ['required', 'integer', 'min:0'],
             'jumps.*.height_cm'  => ['nullable', 'numeric', 'min:0', 'max:150'],
             'jumps.*.type'       => ['nullable', 'string', 'max:32'],
         ]);
